# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  email: string;
  createdAt: string;
}

export interface AuthCredentials {
  /**
     * @minLength 3
     * @maxLength 320
     */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CalculationType = typeof CalculationType[keyof typeof CalculationType];


export const CalculationType = {
  throughput: 'throughput',
  'link-budget': 'link-budget',
} as const;

export interface Calculation {
  id: string;
  type: CalculationType;
  title: string;
  summary: string;
  notes?: string;
  createdAt: string;
}

export interface CreateCalculation {
  type: CalculationType;
  /**
     * @minLength 1
     * @maxLength 255
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 512
     */
  summary: string;
  /** @maxLength 1024 */
  notes?: string;
}