# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and user accounts
  - name: calculations
    description: Saved 5G NR calculations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /calculations:
    get:
      operationId: getCalculations
      tags: [calculations]
      summary: List saved calculations
      description: Returns all saved 5G NR calculations, newest first.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: A list of saved calculations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Calculation"
        "401":
          description: Missing or invalid API key
    post:
      operationId: createCalculation
      tags: [calculations]
      summary: Save a calculation
      description: Stores a new 5G NR calculation result.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalculation"
      responses:
        "201":
          description: The created calculation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calculation"
        "401":
          description: Missing or invalid API key
        "429":
          description: Too many requests
  /calculations/{id}:
    delete:
      operationId: deleteCalculation
      tags: [calculations]
      summary: Delete a calculation
      description: Removes a saved calculation by id.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Missing or invalid API key
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account
      description: Registers a new user with email and password, returning a session token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
        "409":
          description: Email already registered
        "429":
          description: Too many requests
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      description: Authenticates an existing user, returning a session token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
        "401":
          description: Invalid credentials
        "429":
          description: Too many requests
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      description: Revokes the current session token.
      security:
        - ApiKeyAuth: []
      responses:
        "204":
          description: Logged out
        "401":
          description: Missing or invalid token
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Current user
      description: Returns the authenticated user's profile.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: The authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid token
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer token: Authorization: Bearer <API_KEY>"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - createdAt
    AuthCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 128
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    CalculationType:
      type: string
      enum:
        - throughput
        - link-budget
    Calculation:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/CalculationType"
        title:
          type: string
        summary:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - summary
        - createdAt
    CreateCalculation:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/CalculationType"
        title:
          type: string
          minLength: 1
          maxLength: 255
        summary:
          type: string
          minLength: 1
          maxLength: 512
        notes:
          type: string
          maxLength: 1024
      required:
        - type
        - title
        - summary