# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  /**
     * @minLength 3
     * @maxLength 320
     */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}