# lib/db/src/schema/users.ts
import { pgTable, uuid, varchar, text, timestamp } from "drizzle-orm/pg-core";

/** Registered users. Passwords are stored only as scrypt hashes. */
export const usersTable = pgTable("users", {
  id: uuid("id").primaryKey().defaultRandom(),
  email: varchar("email", { length: 320 }).notNull().unique(),
  passwordHash: text("password_hash").notNull(),
  createdAt: timestamp("created_at", { withTimezone: true })
    .notNull()
    .defaultNow(),
});

export type User = typeof usersTable.$inferSelect;
export type InsertUser = typeof usersTable.$inferInsert;